backend/requirements.txt
fastapi
uvicorn
jinja2
httpx
python-dotenv
spotipy
openai